+++ src/lib/Flow/flow.ts
@@ -3,6 +3,16 @@ import { calculateTable, type TableInput, type TableOutput } from './table'
 import { round, roundToPrecision } from '../round'
 import { calculatePerformanceData } from './toLandPerformance'
 import { getClimbRate } from './climbRate'
+import { WB } from '$lib/WBError'
+
+export enum Component {
+	StartAndEnd,
+	Table,
+	PressureAltitude,
+	PerfTemp,
+	Wind,
+	PerfResult
+}
 
 export type CompleteFlowInput = {
 	table: TableInput
@@ -11,6 +21,7 @@ export type CompleteFlowInput = {
 	temperature: number
 	headwind: number
 	toWeightOverride: number
+	climbAlt: number
 }
 
 export type CompleteFlowOutput = {
@@ -21,7 +32,6 @@ export type CompleteFlowOutput = {
 		takeoffRoll: number
 		takeoffFifty: number
 		climbRate: number
-		climbAlt: number
 		landRoll: number
 		landFifty: number
 	}
@@ -31,21 +41,46 @@ export type CompleteFlowOutput = {
 export function flow(input: CompleteFlowInput): CompleteFlowOutput {
 	const calculatedTable = calculateTable(input.table)
 	let maneuveringSpeed
+	// If changing aircraft, use new lnd weight
 	if (!calculatedTable.changeAircraft) {
-		maneuveringSpeed = round(Math.sqrt(calculatedTable.landing.weight / 2550) * 105)
+		maneuveringSpeed = round(Math.sqrt(calculatedTable.landing.weight / 2550) * 105, true)
 	} else {
-		maneuveringSpeed = round(Math.sqrt(calculatedTable.changeAircraft.landing.weight / 2550) * 105)
+		maneuveringSpeed = round(
+			Math.sqrt(calculatedTable.changeAircraft.landing.weight / 2550) * 105,
+			true
+		)
+	}
+	if (input.altimiter > 35 || input.altimiter < 25) {
+		throw new WB(107, 'Invalid altimiter setting detected', Component.PressureAltitude)
 	}
 	const pressureAltitude = roundToPrecision(
 		(29.92 - Number(input.altimiter)) * 1000 + Number(input.fieldElevation),
 		1
 	)
+	if (pressureAltitude > 12000)
+		throw new WB(
+			9999,
+			'Pressure altitude greater than 12,000ft, performance data unavailable',
+			Component.PressureAltitude
+		)
 	let takeoffWeight: number
 	if (!calculatedTable.changeAircraft) {
 		takeoffWeight = calculatedTable.takeoff.weight
 	} else {
-		takeoffWeight = calculatedTable.changeAircraft?.takeoff.weight
+		takeoffWeight = calculatedTable.changeAircraft.takeoff.weight
 	}
+	if (takeoffWeight < 0)
+		throw new WB(
+			9999,
+			'Takeoff weight less than 0lbs, data likely invalid. Cannot continue.',
+			Component.Table
+		)
+	if (takeoffWeight > 10000)
+		throw new WB(
+			9999,
+			'Takeoff weight greater than 10,000lbs, data likely invalid. Cannot continue.',
+			Component.Table
+		)
 	const performanceData = calculatePerformanceData(
 		takeoffWeight,
 		pressureAltitude,
@@ -53,7 +88,7 @@ export function flow(input: CompleteFlowInput): CompleteFlowOutput {
 		input.headwind,
 		input.toWeightOverride
 	)
-	const climbData = getClimbRate(pressureAltitude, input.temperature)
+	const climbData = getClimbRate(input.climbAlt, input.temperature)
 	const limitData = calcLimits(calculatedTable)
 	return {
 		table: calculatedTable,
@@ -61,8 +96,7 @@ export function flow(input: CompleteFlowInput): CompleteFlowOutput {
 		maneuveringSpeed: maneuveringSpeed,
 		performance: {
 			...performanceData,
-			climbAlt: climbData.altitude,
-			climbRate: climbData.rate
+			climbRate: climbData
 		},
 		validation: limitData
 	}